refactor: remove Kani/BLLVM, consolidate benches and docs
- Remove bllvm_optimizations bench, coverage_analysis.sh, estimate_test_coverage*.py
- Remove bllvm_memory_profiling_tests, bllvm_optimization_tests
- Delete obsolete docs: KANI_*, CONSENSUS_COVERAGE_ASSESSMENT, DOCUMENTATION_ORGANIZATION, etc.
- Update benches (hash_operations, mempool_operations), fuzz targets
- Update CI, coverage, CONTRIBUTING, README, SECURITY
- Update docs/PROOF_LIMITATIONS, VERIFICATION, TEST_DATA_SOURCES

Made-with: Cursor

+++ src/orange_paper_constants.rs
@@ -9,62 +9,61 @@
 //! Each constant is linked to its Orange Paper section via documentation comments.
 
 /// satoshis per BTC, see [Economic Model](#6-economic-model)
-/// 
+///
 /// Source: Orange Paper Section 4.1
 /// Formula: $C = 10^8$
 pub const C: u64 = 100_000_000;
 
 /// maximum money supply, see [Supply Limit](#supply-limit)
-/// 
+///
 /// Source: Orange Paper Section 4.1
 /// Formula: $M_MAX = 21 \times 10^6 \times C$
-pub const M_MAX: i64 = ((21_000_000 * C) as i64) as i64;
+pub const M_MAX: i64 = (21_000_000 * C) as i64;
 
 /// halving interval, see [Block Subsidy](#61-block-subsidy)
-/// 
+///
 /// Source: Orange Paper Section 4.1
 /// Formula: $H = 210,000$
 pub const H: u64 = 210_000;
 
 /// maximum block weight, see [Block Validation](#53-block-validation)
-/// 
+///
 /// Source: Orange Paper Section 4.2
 /// Formula: $W_MAX = 4 \times 10^6$
 pub const W_MAX: u64 = 4_000_000;
 
 /// maximum sigops per block, see [Script Execution](#52-script-execution)
-/// 
+///
 /// Source: Orange Paper Section 4.2
 /// Formula: $S_MAX = 80,000$
 pub const S_MAX: u64 = 80_000;
 
 /// coinbase maturity requirement, see [Transaction Validation](#51-transaction-validation)
-/// 
+///
 /// Source: Orange Paper Section 4.2
 /// Formula: $R = 100$
 pub const R: u64 = 100;
 
 /// maximum script length, see [Script Security](#script-security)
-/// 
+///
 /// Source: Orange Paper Section 4.3
 /// Formula: $L_SCRIPT = 10,000$
 pub const L_SCRIPT: u64 = 10_000;
 
 /// maximum stack size, see [Script Execution Bounds](#theorem-84)
-/// 
+///
 /// Source: Orange Paper Section 4.3
 /// Formula: $L_STACK = 1,000$
 pub const L_STACK: u64 = 1_000;
 
 /// maximum operations per script, see [Script Execution Bounds](#theorem-84)
-/// 
+///
 /// Source: Orange Paper Section 4.3
 /// Formula: $L_OPS = 201$
 pub const L_OPS: u64 = 201;
 
 /// maximum element size, see [Script Execution](#52-script-execution)
-/// 
+///
 /// Source: Orange Paper Section 4.3
 /// Formula: $L_ELEMENT = 520$
 pub const L_ELEMENT: u64 = 520;
-
